fix pattern for llm logs table
